Day 24
There is way more solution than there needs to be. The VM is cool, so wanted to leave it in as a way to verify the solution, even though it is unnecessary. This 100% was a day of just needing to analyze the input file and write a specific solution by hand.